+++ tests/subsys/fs/nffs_fs_api/basic/nrf5x.conf
