+++ tests/subsys/fs/nffs_fs_api/basic/prj.conf
