+++ tests/subsys/fs/nffs_fs_api/basic/src/main.c
@@ -7,56 +7,8 @@
 #include <nffs/nffs.h>
 #include "test_nffs.h"
 
-#if CONFIG_BOARD_QEMU_X86
-static struct nffs_area_desc nffs_selftest_area_descs[] = {
-	{ 0x00000000, 16 * 1024 },
-	{ 0x00004000, 16 * 1024 },
-	{ 0x00008000, 16 * 1024 },
-	{ 0x0000c000, 16 * 1024 },
-	{ 0x00010000, 64 * 1024 },
-	{ 0x00020000, 128 * 1024 },
-	{ 0x00040000, 128 * 1024 },
-	{ 0x00060000, 128 * 1024 },
-	{ 0x00080000, 128 * 1024 },
-	{ 0x000a0000, 128 * 1024 },
-	{ 0x000c0000, 128 * 1024 },
-	{ 0x000e0000, 128 * 1024 },
-	{ 0, 0 },
-};
-#else
-static struct nffs_area_desc nffs_selftest_area_descs[] = {
-	{ 0x00020000, 2 * 4096 },
-	{ 0x00022000, 2 * 4096 },
-	{ 0x00024000, 2 * 4096 },
-	{ 0x00026000, 2 * 4096 },
-	{ 0x00028000, 2 * 4096 },
-	{ 0x0002a000, 2 * 4096 },
-	{ 0x0002c000, 2 * 4096 },
-	{ 0x00030000, 2 * 4096 },
-	{ 0x00032000, 2 * 4096 },
-	{ 0x00034000, 2 * 4096 },
-	{ 0x00036000, 2 * 4096 },
-	{ 0x00038000, 2 * 4096 },
-	{ 0, 0 },
-};
-#endif
-
-static struct nffs_area_desc *save_area_descs;
-
-static void test_setup(void)
-{
-	save_area_descs = nffs_current_area_descs;
-	nffs_current_area_descs = nffs_selftest_area_descs;
-}
-
-static void test_teardown(void)
-{
-	nffs_current_area_descs = save_area_descs;
-}
-
 void test_main(void)
 {
-#ifdef TEST_basic
 	ztest_test_suite(nffs_fs_basic_test,
 		ztest_unit_test_setup_teardown(test_fs_mount,
 					       test_setup, test_teardown),
@@ -105,34 +57,4 @@ void test_main(void)
 		ztest_unit_test_setup_teardown(test_gc_on_oom,
 					       test_setup, test_teardown));
 	ztest_run_test_suite(nffs_fs_basic_test);
-#endif
-
-#ifdef TEST_large
-	ztest_test_suite(nffs_fs_large_test,
-		ztest_unit_test_setup_teardown(test_fs_mount,
-					       test_setup, test_teardown),
-		ztest_unit_test_setup_teardown(test_large_unlink,
-					       test_setup, test_teardown),
-		ztest_unit_test_setup_teardown(test_large_system,
-					       test_setup, test_teardown));
-	ztest_run_test_suite(nffs_fs_large_test);
-#endif
-
-#ifdef TEST_cache
-	ztest_test_suite(nffs_fs_cache_test,
-		ztest_unit_test_setup_teardown(test_fs_mount,
-					       test_setup, test_teardown),
-		ztest_unit_test_setup_teardown(test_cache_large_file,
-					       test_setup, test_teardown));
-	ztest_run_test_suite(nffs_fs_cache_test);
-#endif
-
-#ifdef TEST_performance
-	ztest_test_suite(nffs_fs_performace_test,
-		ztest_unit_test_setup_teardown(test_fs_mount,
-					       test_setup, test_teardown),
-		ztest_unit_test_setup_teardown(test_performance,
-					       test_setup, test_teardown));
-	ztest_run_test_suite(nffs_fs_performace_test);
-#endif
 }
