+++ tests/subsys/fs/nffs_fs_api/basic/testcase.yaml
@@ -1,0 +1,6 @@
+common:
+  tags: nffs filesystem
+tests:
+  filesystem.nffs.basic:
+    extra_args: TEST=basic
+    platform_whitelist: qemu_x86
