+++ tests/subsys/fs/nffs_fs_api/common/nffs_test_utils.h
