+++ tests/subsys/fs/nffs_fs_api/common/test_corrupt_block.c
