+++ tests/subsys/fs/nffs_fs_api/common/test_corrupt_scratch.c
