+++ tests/subsys/fs/nffs_fs_api/large/nrf5x.conf
@@ -1,0 +1,20 @@
+CONFIG_ZTEST=y
+CONFIG_ZTEST_STACKSIZE=2048
+CONFIG_MAIN_STACK_SIZE=1024
+CONFIG_HEAP_MEM_POOL_SIZE=1024
+
+CONFIG_FLASH=y
+CONFIG_SOC_FLASH_NRF_RADIO_SYNC=y
+CONFIG_MPU_ALLOW_FLASH_WRITE=y
+
+CONFIG_FILE_SYSTEM=y
+CONFIG_FILE_SYSTEM_NFFS=y
+CONFIG_FS_NFFS_FLASH_DEV_NAME="NRF_FLASH_DRV_NAME"
+CONFIG_FS_NFFS_NUM_FILES=4
+CONFIG_FS_NFFS_NUM_DIRS=4
+CONFIG_FS_NFFS_NUM_INODES=1024
+CONFIG_FS_NFFS_NUM_BLOCKS=1024
+CONFIG_FS_NFFS_NUM_CACHE_INODES=1
+CONFIG_FS_NFFS_NUM_CACHE_BLOCKS=1
+CONFIG_FILE_SYSTEM_NFFS=y
+CONFIG_NFFS_FILESYSTEM_MAX_AREAS=12
