+++ tests/subsys/fs/nffs_fs_api/large/prj.conf
@@ -1,0 +1,17 @@
+CONFIG_TEST_FLASH_DRIVERS=y
+CONFIG_FILE_SYSTEM_NFFS=y
+CONFIG_FILE_SYSTEM=y
+CONFIG_FLASH_PAGE_LAYOUT=y
+CONFIG_FLASH=y
+CONFIG_FS_NFFS_FLASH_DEV_NAME="ram_flash_test_drv"
+CONFIG_FS_NFFS_NUM_BLOCKS=1024
+CONFIG_FS_NFFS_NUM_CACHE_BLOCKS=1
+CONFIG_FS_NFFS_NUM_CACHE_INODES=1
+CONFIG_FS_NFFS_NUM_DIRS=4
+CONFIG_FS_NFFS_NUM_FILES=4
+CONFIG_FS_NFFS_NUM_INODES=1024
+CONFIG_HEAP_MEM_POOL_SIZE=1024
+CONFIG_MAIN_STACK_SIZE=1024
+CONFIG_NFFS_FILESYSTEM_MAX_AREAS=12
+CONFIG_ZTEST_STACKSIZE=2048
+CONFIG_ZTEST=y
