+++ tests/subsys/fs/nffs_fs_api/large/testcase.yaml
@@ -1,0 +1,6 @@
+common:
+  tags: nffs filesystem
+tests:
+  filesystem.nffs.large:
+    extra_args: TEST=large
+    platform_whitelist: qemu_x86
