--- tests/subsys/fs/nffs_fs_api/testcase.yaml
@@ -1,18 +1,0 @@
-common:
-  tags: nffs filesystem
-tests:
-  filesystem.nffs.basic:
-    extra_args: TEST=basic
-    platform_whitelist: qemu_x86
-  filesystem.nffs.cache:
-    extra_args: TEST=cache
-    extra_configs:
-      - CONFIG_FS_NFFS_NUM_CACHE_BLOCKS=64
-      - CONFIG_FS_NFFS_NUM_CACHE_INODES=4
-    platform_whitelist: qemu_x86
-  filesystem.nffs.large:
-    extra_args: TEST=large
-    platform_whitelist: qemu_x86
-  filesystem.nffs.performance:
-    extra_args: TEST=performance
-    platform_whitelist: qemu_x86
